fix: tests & rename
