+++ index.js
@@ -1,4 +1,4 @@
 module.exports = {
     Broker: require('./broker'),
-    Bus: require('./bus')
+    ServiceBus: require('./serviceBus')
 };
