--- test/cover.js
@@ -1,2 +1,0 @@
-var glob = require('glob');
-glob.sync('**/test*.js', {cwd: __dirname}).forEach((test) => require('./' + test));
